Adds the ability to search youtube videos that have been submitted to reddit.  Obviously someone can also just search through the youtube channel for videos, but I thought that this might be an interesting way to find videos that were liked a lot by reddit users, instead of just a blind YouTube search.
I'm also about ready to get rid of the categories file completely.  I'm now maintaining a generic list of subreddits through a multireddit on Reddit and I also have a gaming subreddit.  This will be easier to maintain instead of just doing a pull request every time I want to add a subreddit to a category.